chore(suite): move onboarding reducer
# Conflicts:
#	packages/suite/src/middlewares/onboarding/__tests__/onboardingMiddleware.test.ts
#	packages/suite/src/support/tests/__fixtures__/defaultAppState.ts

# Conflicts:
#	packages/suite/src/actions/onboarding/onboardingActions.ts
#	packages/suite/src/components/onboarding/SkipStepConfirmation.tsx
#	packages/suite/src/config/onboarding/steps.ts
#	packages/suite/src/hooks/suite/useOnboarding.ts
#	packages/suite/src/middlewares/onboarding/__tests__/onboardingMiddleware.test.ts
#	packages/suite/src/reducers/store.ts
#	packages/suite/src/support/tests/__fixtures__/defaultAppState.ts
#	packages/suite/src/types/onboarding/index.ts
#	packages/suite/src/utils/onboarding/__tests__/steps.test.ts
#	packages/suite/src/utils/onboarding/steps.ts
#	packages/suite/src/views/onboarding/index.tsx
#	packages/suite/src/views/onboarding/steps/DeviceAuthenticityStep/SecurityCheck.tsx
#	suite/onboarding/src/onboardingReducer.ts

# Conflicts:
#	packages/suite/package.json
#	packages/suite/src/views/onboarding/steps/CreateOrRecoverStep.tsx
#	packages/suite/src/views/onboarding/steps/ResetDeviceStep.tsx
#	yarn.lock
