--- packages/suite/src/actions/onboarding/constants/index.ts
@@ -1,3 +1,0 @@
-import * as ONBOARDING from './onboarding';
-
-export { ONBOARDING };
