+++ packages/suite/src/components/onboarding/SkipStepConfirmation.tsx
@@ -1,11 +1,11 @@
 import { type ReactNode } from 'react';
 
 import { Translation } from '@suite/intl';
+import { type AnyStepId, STEP } from '@suite/onboarding';
 import { Modal } from '@trezor/components';
 
-import * as STEP from 'src/constants/onboarding/steps';
+
 import { useOnboarding } from 'src/hooks/suite';
-import { type AnyStepId } from 'src/types/onboarding';
 
 type SkipStepConfirmationProps = {
     onCancel: () => void;
