--- packages/suite/src/constants/onboarding/index.ts
@@ -1,3 +1,0 @@
-import * as STEPS from './steps';
-
-export { STEPS };
