--- packages/suite/src/reducers/onboarding/index.ts
@@ -1,3 +1,0 @@
-import onboarding from 'src/reducers/onboarding/onboardingReducer';
-
-export default onboarding;
