+++ packages/suite/src/support/tests/__fixtures__/defaultAppState.ts
@@ -1,6 +1,7 @@
 import type { BackupState } from '@suite/backup';
 import { initialState as experimentalFeedbackInitialState } from '@suite/experimental-feedback';
 import { flagsInitialState } from '@suite/flags';
+import { type OnboardingState } from '@suite/onboarding';
 import { locksInitialState } from '@suite/locks';
 import { type RouterState } from '@suite/router';
 import { suiteSettingsInitialState } from '@suite/settings';
@@ -12,7 +13,6 @@ import { type NetworkSymbol } from '@suite-common/wallet-config';
 
 import { initialState } from 'src/actions/device/deviceSlice';
 import { initialSuiteSyncDesktopState } from 'src/actions/suiteSync/suiteSyncSlice';
-import { type OnboardingState } from 'src/reducers/onboarding/onboardingReducer';
 import { type AppState } from 'src/reducers/store';
 import { desktopUpdateInitialState } from 'src/reducers/suite/desktopUpdateReducer';
 import { type ProtocolState } from 'src/reducers/suite/protocolReducer';
