+++ packages/suite/src/utils/onboarding/__tests__/steps.test.ts
@@ -1,7 +1,7 @@
+import { STEP } from '@suite/onboarding';
 import { type AcquiredDevice } from '@suite-common/suite-types';
 import { DeviceModelInternal, FirmwareType } from '@trezor/device-utils';
 
-import * as STEP from 'src/constants/onboarding/steps';
 import { type Step, type StepCategory } from 'src/types/onboarding';
 
 import {
