+++ packages/suite/src/views/onboarding/index.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useMemo } from 'react';
 
 import { goto } from '@suite/router';
+import { STEP } from '@suite/onboarding';
 import { selectSelectedDevice } from '@suite-common/device';
 import { selectThpStep } from '@suite-common/thp';
 import { exhaustive } from '@trezor/type-utils';
 
 import { OnboardingLayout } from 'src/components/onboarding/OnboardingLayout';
-import * as STEP from 'src/constants/onboarding/steps';
 import { useDispatch, useOnboarding, useSelector } from 'src/hooks/suite';
 import { UnexpectedState } from 'src/views/onboarding/UnexpectedState';
 import { BackupStep } from 'src/views/onboarding/steps/BackupStep';
