+++ suite/onboarding/package.json
@@ -1,0 +1,21 @@
+{
+    "name": "@suite/onboarding",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "depcheck": "yarn g:depcheck",
+        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
+        "test:unit": "yarn g:jest",
+        "type-check": "yarn g:tsc --build"
+    },
+    "dependencies": {
+        "@reduxjs/toolkit": "2.11.2",
+        "@suite-common/suite-types": "workspace:*",
+        "@suite/analytics": "workspace:*",
+        "@trezor/connect": "workspace:*",
+        "immer": "11.0.0"
+    }
+}
