+++ suite/onboarding/src/onboardingConstants.ts
