+++ suite/onboarding/src/onboardingSteps.ts
