+++ suite/onboarding/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [
+        {
+            "path": "../../suite-common/suite-types"
+        },
+        { "path": "../analytics" },
+        { "path": "../../packages/connect" }
+    ]
+}
